requirements.txt
# EEG音乐系统服务器依赖包
openai>=0.27.0
pygame>=2.1.0
pyserial>=3.5
torcheeg>=1.0.11
moabb>=0.4.6
numpy>=1.21.0
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
eventlet==0.33.3

# 异步HTTP
aiohttp>=3.8.0
websockets>=10.0

# 机器学习框架
torch>=1.13.0

# 数据科学和分析
scikit-learn>=1.1.0
pandas>=1.5.0
seaborn>=0.11.0

# 其他工具
tqdm>=4.64.0

# 音频处理
matplotlib>=3.5.0 